Moved InspectTask to its own class
