+++ src/main/java/com/cyprias/chunkspawnerlimiter/Plugin.java
@@ -1,11 +1,9 @@
 package com.cyprias.chunkspawnerlimiter;
 
 import java.io.File;
-import java.util.HashMap;
 import java.util.logging.Logger;
 
 import org.bstats.bukkit.Metrics;
-import org.bukkit.Location;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.configuration.file.YamlConfiguration;
 import org.bukkit.plugin.java.JavaPlugin;
